requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-openai==0.0.5
python-dotenv==1.0.0

# HubSpot Integration
hubspot-api-client==12.0.0

# Ad Platform Integrations
google-ads==28.4.0
facebook-business==21.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Database (for storing attribution data)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1
faker==22.0.0

# Supabase for RAG
supabase==2.3.0
vecs==0.4.0

# RAG & Vector Store
tiktoken==0.5.2
openai==1.12.0

# Error Handling
tenacity==8.2.3

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Secrets Management
boto3==1.34.34
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0

# Rate Limiting
ratelimit==2.2.1

# Caching
redis==5.0.1
cachetools==5.3.2
